+++ Source/VagabondsWork/MarkerComponent.cpp
@@ -1,6 +1,13 @@
 #include "MarkerComponent.h"
+#include "Engine/GameInstance.h"
 
 UMarkerComponent::UMarkerComponent()
 {
 	PrimaryComponentTick.bCanEverTick = false;
+	Faction = EFaction::VoidVagabonds;
+}
+
+void UMarkerComponent::BeginPlay()
+{
+	Super::BeginPlay();
 }