+++ Source/VagabondsWork/MarkerComponent.h
@@ -2,6 +2,7 @@
 
 #include "CoreMinimal.h"
 #include "Components/ActorComponent.h"
+#include "FactionsSubsystem.h"
 #include "MarkerComponent.generated.h"
 
 UENUM(BlueprintType)
@@ -13,17 +14,6 @@ enum class EMarkerType : uint8
 	Component UMETA(DisplayName = "Component")
 };
 
-UENUM(BlueprintType)
-enum class EFaction : uint8
-{
-	VoidVagabonds UMETA(DisplayName = "Void Vagabonds"),
-	VoidRaiders UMETA(DisplayName = "Void Raiders"),
-	SilentMandate UMETA(DisplayName = "Silent Mandate"),
-	FederationOfNations UMETA(DisplayName = "Federation of Nations"),
-	UnitedRepublic UMETA(DisplayName = "United Republic"),
-	ConcordUnion UMETA(DisplayName = "Concord Union")
-};
-
 UCLASS(ClassGroup=(Custom), meta=(BlueprintSpawnableComponent))
 class VAGABONDSWORK_API UMarkerComponent : public UActorComponent
 {
@@ -31,9 +21,10 @@ class VAGABONDSWORK_API UMarkerComponent : public UActorComponent
 
 public:
 	UMarkerComponent();
+	virtual void BeginPlay() override;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Marker")
 	EMarkerType MarkerType = EMarkerType::Component;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Faction")
-	EFaction Faction = EFaction::VoidVagabonds;
+	EFaction Faction;
 };